frontend: Move LightTooltip to the common list of components
