+++ frontend/src/js/components/Common/LightTooltip.tsx
@@ -1,0 +1,12 @@
+import { Tooltip, withStyles } from '@material-ui/core';
+
+const LightTooltip = withStyles(theme => ({
+  tooltip: {
+    backgroundColor: theme.palette.common.white,
+    color: 'rgba(0, 0, 0, 0.87)',
+    boxShadow: theme.shadows[1],
+    fontSize: '1rem',
+    whiteSpace: 'pre-line',
+  },
+}))(Tooltip);
+export default LightTooltip;
